chore(map): adjust hashCode and remove loadFactor
 * usage of loadFactor is removed from the map -
 we will use overflow only if the max hash table size is reached
 -> better performance
 * improve hash function to distribute entries better
 -> add test case to check the growing of the hash table
 * change hashcode from long to int again

 related to zeebe-io/zeebe#576

+++ zb-map/src/main/java/io/zeebe/map/KeyHandler.java
@@ -19,7 +19,7 @@ public interface KeyHandler
 {
     void setKeyLength(int keyLength);
 
-    long keyHashCode();
+    int keyHashCode();
 
     boolean keyEquals(long keyAddr);
